+++ tests/testthat/test-relpath.R
@@ -8,7 +8,8 @@ add_file("script/file.txt", "something")
 add_file("data/filespec.txt", "something")
 logCreate()
 
-test_that("relpath outputs the relative path when given an unresolved or absolute path", {
+# Confirm users are able to provide a path with ../ or give an absolute path
+test_that("relpath outputs script paths relative to the QC log location: for unresolved or absolute path [REV-REL-001]", {
   expect_identical("script/file.txt", pathFromLogRoot("data/../script/file.txt"))
   expect_identical("script/file.txt", relPath("data/../script/file.txt"))
   expect_identical("script/file.txt", pathFromLogRoot("/tmp/svn-testing/test/script/file.txt"))
@@ -20,7 +21,7 @@ logAssign("script/../data/filespec.txt")
 add_commit("first")
 qclog <- readr::read_csv("QClog.csv")
 
-test_that("Only resolved file paths are stored in the QClog", {
+test_that("relpath outputs script paths relative to the QC log location - only resolved file paths are stored in the QClog [REV-REL-001]", {
   expect_true(length(unique(qclog$file)) == 2)
 })
 
@@ -30,7 +31,8 @@ qclog_old <- qclog %>% dplyr::slice(1) %>% dplyr::mutate(file = "data/../script/
 qclog_combine <- qclog %>% dplyr::bind_rows(qclog_old)
 readr::write_csv(qclog_combine, "QClog.csv")
 
-test_that("logPending only includes resolved paths when unresolved paths present", {
+# Confirm legacy QClogs will not show duplicate records for same file with different paths
+test_that("logPending only includes resolved paths when unresolved paths present [REV-PND-002]", {
   pendingcsv <- logPending()
   expect_true(length(unique(pendingcsv$file)) == 2)
 })